Fix the usage of internal hal ids for all data structures (#2033)
* Commit partial changes

* Changes based on review comments

* Fix compilation error

* Fix compilation issue

+++ nic/hal/src/proxy.hpp
@@ -70,8 +70,8 @@ enum {
 };
 
 typedef uint32_t lif_id_t;
-typedef uint8_t  qtype_t;
-typedef uint32_t  qid_t;
+typedef uint32_t  qtype_t;
+typedef uint64_t  qid_t;
 
 typedef struct proxy_meta_qtype_s {
     qtype_t     qtype_val;
